backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
PyMuPDF
pillow==10.4.0
python-multipart==0.0.9
pydantic==2.9.2
aiofiles==24.1.0
reportlab==4.2.2
google-genai>=1.0.0
python-dotenv>=1.0.0
